Extract implementation to different file
