+++ TwoPaneLayout/library/src/main/java/com/microsoft/device/dualscreen/twopanelayout/Screen.kt
@@ -4,6 +4,10 @@ package com.microsoft.device.dualscreen.twopanelayout
  * Class that represents the screens in the NavHost
  */
 sealed class Screen(val route: String) {
+    override fun toString(): String {
+        return route
+    }
+
     /**
      * Screen object representing pane 1
      */
