+++ TwoPaneLayout/library/src/main/java/com/microsoft/device/dualscreen/twopanelayout/TwoPaneLayoutNavCore.kt
@@ -1,0 +1,101 @@
+package com.microsoft.device.dualscreen.twopanelayout
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.geometry.Size
+import androidx.compose.ui.layout.Layout
+import androidx.compose.ui.platform.LocalDensity
+import androidx.navigation.NavHostController
+import androidx.navigation.NavOptionsBuilder
+import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
+import com.microsoft.device.dualscreen.windowstate.WindowState
+
+internal var isSinglePaneLayout = true
+private var currentPane1 = mutableStateOf("")
+private var currentPane2 = mutableStateOf("")
+private var currentSinglePane = mutableStateOf("")
+
+internal var navigatePane1To: NavHostController.(String) -> Unit = { _: String -> }
+internal var navigatePane2To: NavHostController.(String) -> Unit = { _: String -> }
+
+internal fun NavHostController.navigateSinglePaneTo(route: String, navOptions: NavOptionsBuilder.() -> Unit) {
+    navigate(route, navOptions)
+    currentSinglePane.value = route
+}
+
+private fun findDestination(route: String, destinations: Array<Destination>): Destination {
+    return destinations.find { pane -> pane.route == route }
+        ?: throw IllegalArgumentException("Invalid route $route, not present in list of destinations $destinations")
+}
+
+
+@Composable
+internal fun SinglePaneContainer(
+    destinations: Array<Destination>,
+    startDestination: String,
+    navController: NavHostController,
+) {
+    currentSinglePane.value = startDestination
+
+    NavHost(
+        navController = navController,
+        startDestination = startDestination
+    ) {
+        destinations.forEach { pane ->
+            composable(pane.route) {
+                TwoPaneNavScopeInstance.(pane.content)()
+            }
+        }
+    }
+}
+
+@Composable
+internal fun TwoPaneContainer(
+    windowState: WindowState,
+    modifier: Modifier,
+    destinations: Array<Destination>,
+    pane1StartDestination: String,
+    pane2StartDestination: String
+) {
+    // Calculate pane sizes
+    val pane1SizePx: Size
+    val pane2SizePx: Size
+    with(LocalDensity.current) {
+        pane1SizePx = windowState.pane1SizeDp.toSize()
+        pane2SizePx = windowState.pane2SizeDp.toSize()
+    }
+
+    // Initialize start destinations
+    currentPane1.value = pane1StartDestination
+    currentPane2.value = pane2StartDestination
+
+    // Initialize navigation method handlers
+    navigatePane1To = { route ->
+        findDestination(route, destinations)
+        currentPane1.value = route
+    }
+    navigatePane2To = { route ->
+        findDestination(route, destinations)
+        currentPane2.value = route
+    }
+
+    // Find the destinations to display in each pane
+    val pane1 = findDestination(currentPane1.value, destinations).content
+    val pane2 = findDestination(currentPane2.value, destinations).content
+
+    val measurePolicy = twoPaneMeasurePolicy(
+        windowMode = windowState.windowMode,
+        isSeparating = windowState.foldIsSeparating,
+        paneSizes = arrayOf(pane1SizePx, pane2SizePx),
+    )
+    Layout(
+        content = {
+            TwoPaneNavScopeInstance.pane1()
+            TwoPaneNavScopeInstance.pane2()
+        },
+        measurePolicy = measurePolicy,
+        modifier = modifier
+    )
+}
